requirements.txt
aniso8601==9.0.1
attrs==24.2.0
bidict==0.23.1
blinker==1.8.2
boto3==1.35.9
boto3-stubs==1.35.9
botocore==1.35.9
botocore-stubs==1.35.9
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
ecdsa==0.19.0
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
fsspec==2024.6.1
future==1.0.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.14.0
huggingface-hub==0.24.6
idna==3.8
itsdangerous==2.2.0
jamotools==0.1.10
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mpmath==1.3.0
mypy-boto3-cognito-idp==1.35.0
networkx==3.3
numpy==2.1.0
packaging==24.1
promise==2.3
pyasn1==0.6.0
PyJWT==2.9.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-jose==3.3.0
python-mecab-ko==1.3.7
python-mecab-ko-dic==2.1.1.post2
python-socketio==5.11.3
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rpds-py==0.20.0
rsa==4.9
Rx==1.6.3
s3transfer==0.10.2
safetensors==0.4.4
setuptools==74.0.0
simple-websocket==1.0.0
six==1.16.0
sympy==1.13.2
tokenizers==0.19.1
tqdm==4.66.5
transformers==4.44.2
types-awscrt==0.21.2
types-s3transfer==0.10.2
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
wsproto==1.2.0
